requirements.txt
django
django-environ
django-ckeditor
daphne
Twisted[tls,http2]
mysqlclient
